Move tests that just test StaticHeader behavior into that module
For more precision and better encapsulation.
Rename or improve the test documentation as appropriate.
